Move helper functions into MasternodeHelpers
