+++ divi/src/masternode.cpp
@@ -17,7 +17,6 @@
 #include <wallet.h>
 #include <utiltime.h>
 #include <WalletTx.h>
-#include <MasternodeHelpers.h>
 #include <script/standard.h>
 #include <blockmap.h>
 #include <chainparams.h>
@@ -41,106 +40,6 @@ static CAmount getCollateralAmount(MasternodeTier tier)
     return static_cast<CAmount>(-1.0);
   }
 }
-
-bool GetBlockHashForScoring(uint256& hash, int nBlockHeight)
-{
-    const auto* tip = chainActive.Tip();
-    if (tip == nullptr)
-        return false;
-    return GetBlockHashForScoring(hash, tip, nBlockHeight - tip->nHeight);
-}
-
-bool GetBlockHashForScoring(uint256& hash, const CBlockIndex* pindex, const int offset)
-{
-    if (pindex == nullptr)
-        return false;
-
-    const auto* pindexAncestor = pindex->GetAncestor(pindex->nHeight + offset - 101);
-    if (pindexAncestor == nullptr)
-        return false;
-
-    hash = pindexAncestor->GetBlockHash();
-    return true;
-}
-
-const CBlockIndex* ComputeCollateralBlockIndex(const CMasternode& masternode)
-{
-    static std::map<COutPoint,const CBlockIndex*> cachedCollateralBlockIndices;
-
-    const CBlockIndex* collateralBlockIndex = cachedCollateralBlockIndices[masternode.vin.prevout];
-    if (collateralBlockIndex)
-    {
-        if(chainActive.Contains(collateralBlockIndex))
-        {
-            return collateralBlockIndex;
-        }
-    }
-
-    uint256 hashBlock;
-    CTransaction tx;
-    if (!GetTransaction(masternode.vin.prevout.hash, tx, hashBlock, true)) {
-        collateralBlockIndex = nullptr;
-        return collateralBlockIndex;
-    }
-
-    const auto mi = mapBlockIndex.find(hashBlock);
-    if (mi == mapBlockIndex.end() || mi->second == nullptr) {
-        collateralBlockIndex = nullptr;
-        return collateralBlockIndex;
-    }
-
-    if (!chainActive.Contains(mi->second)) {
-        collateralBlockIndex = nullptr;
-        return collateralBlockIndex;
-    }
-    collateralBlockIndex = mi->second;
-    return collateralBlockIndex;
-}
-
-const CBlockIndex* ComputeMasternodeConfirmationBlockIndex(const CMasternode& masternode)
-{
-    const CBlockIndex* pindexConf = nullptr;
-    {
-        LOCK(cs_main);
-        const auto* pindexCollateral = ComputeCollateralBlockIndex(masternode);
-        if (pindexCollateral == nullptr)
-            pindexConf = nullptr;
-        else {
-            assert(chainActive.Contains(pindexCollateral));
-            pindexConf = chainActive[pindexCollateral->nHeight + MASTERNODE_MIN_CONFIRMATIONS - 1];
-            assert(pindexConf == nullptr || pindexConf->GetAncestor(pindexCollateral->nHeight) == pindexCollateral);
-        }
-    }
-    return pindexConf;
-}
-
-int ComputeMasternodeInputAge(const CMasternode& masternode)
-{
-    LOCK(cs_main);
-
-    const auto* pindex = ComputeCollateralBlockIndex(masternode);
-    if (pindex == nullptr)
-        return 0;
-
-    assert(chainActive.Contains(pindex));
-
-    const unsigned tipHeight = chainActive.Height();
-    assert(tipHeight >= pindex->nHeight);
-
-    return tipHeight - pindex->nHeight + 1;
-}
-
-CMasternodePing createCurrentPing(const CTxIn& newVin)
-{
-    CMasternodePing ping;
-    ping.vin = newVin;
-    ping.blockHash = chainActive[chainActive.Height() - 12]->GetBlockHash();
-    ping.sigTime = GetAdjustedTime();
-    ping.signature = std::vector<unsigned char>();
-    return ping;
-}
-
-
 
 CAmount CMasternode::GetTierCollateralAmount(const MasternodeTier tier)
 {
